+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 * `Game` trait
+* `Playdate` struct
 * `game_loop!` macro
 
 [Unreleased]: https://github.com/jcornaz/beancount_parser_2/compare/...HEAD